refactor: make enter leave logic in custom hook by component life cycle
