+++ frontend/src/domains/admin/pages/chat-page/index.tsx
@@ -13,13 +13,13 @@ import {
   SendChatData,
 } from './ChatData.tsx';
 import { useNavigate, useParams } from 'react-router-dom';
-import { forwardRef, useCallback, useState, useEffect, useRef } from 'react';
+import { forwardRef, useCallback, useState, useEffect } from 'react';
 // import { getDefaultBorderStyle } from '@/components/border/getBorderType.tsx';
 // import { BorderType } from '@/components/border/BorderProps.tsx';
-import { useSocketManager } from '@/hooks/useSocketManager';
 import { useSocketStore, ChatMessage } from '@/store/socketStore';
 import { useImageUpload } from '@/hooks/useImageUpload';
 import { ImageFileSizeDialog } from '@/components/Dialog/ImageFileSizeDialog.tsx';
+import { useEnterLeaveHandler } from '@/hooks/useEnterLeaveHandler.ts';
 
 interface ChatPageProps {
   apiChatData: ChatData[];
@@ -59,7 +59,6 @@ const ChatPage = forwardRef<HTMLDivElement, ChatPageProps>(
     let upLastItemId: string = '';
     let downLatItemId: string = '';
 
-    const socketManager = useSocketManager();
     const { subscribe, sendMessage } = useSocketStore();
 
     const handleMessageReceive = useCallback(
@@ -81,20 +80,7 @@ const ChatPage = forwardRef<HTMLDivElement, ChatPageProps>(
       [roomId, memberId],
     );
 
-    const hasLeft = useRef(false);
-
-    const handleLeave = useCallback(() => {
-      if (!hasLeft.current) {
-        socketManager('AGENDA', 'LEAVE', Number(localStorage.getItem('member_id')), 'ADMIN');
-        hasLeft.current = true;
-      }
-    }, [socketManager]);
-
-    useEffect(() => {
-      return () => {
-        handleLeave();
-      };
-    }, [handleLeave]);
+    useEnterLeaveHandler('AGENDA', 'ADMIN');
 
     useEffect(() => {
       const unsubscribe = subscribe('AGENDA', Number(roomId), handleMessageReceive);
@@ -117,7 +103,6 @@ const ChatPage = forwardRef<HTMLDivElement, ChatPageProps>(
           title={chatRoomInfo.title}
           rightIconSrc="/src/assets/icons/information-circle-contained.svg"
           onLeftIconClick={() => {
-            handleLeave();
             navigate(-1);
           }}
           onRightIconClick={() => {}}
